Restored packages support.

+++ Models/DTOs/Package/PublicPackage.cs
@@ -2,10 +2,10 @@
 {
     public class PublicPackage
     {
-        public required uint Id { get; set; }
+        public required ulong Id { get; set; }
         public required DateTimeOffset CreatedAt { get; set; }
         public required DateTimeOffset UpdatedAt { get; set; }
-        public required uint ProductId { get; set; }
+        public required ulong ProductId { get; set; }
         public required ulong FileSize { get; set; }
         public required string Name { get; set; }
         public required string Version { get; set; }
